Fix dhcpcd service
First of all, fix the dependency name in the per-interface service file.

Second, we allow the master services daemonize itself and change the target
to wait. This way, we block init until we are sure the dhcpcd master
is running. Otherwise we have a race condition if init gets scheduled
for some reason and immediately starts the dhcpcd interface daemon that
cannot see the master and decides that *it* is the msater now.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@tele2.at>