Cleanup DHCP client/server configuration
Make sure dhcpcd does not interfere with resolv.conf. For the router
product we have a dnsmasq running that forwards to a local root
resolver (i.e. unbound).

For the ALIX board, we want to run a DHCP client on one port and a
DHCP server on the other two.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@tele2.at>