Downgrade binutils to 2.30
This is a workaround for an issue with binutils on statically linked x86
binaries. The tls section has some loading flag set wrong and is not mapped
into memory by the kernel. The musl startup code tries to parse the section
information and segfaults.

See: http://www.openwall.com/lists/musl/2018/07/18/5

Signed-off-by: David Oberhollenzer <david.oberhollenzer@tele2.at>